Updated to get Google accessToken using Javascript on robot.
We removed the need to set up and use a Google Cloud Function to get a Google Access Token.  A signed JWT is now sent in a request directly in Javascript on the robot. This removes the security hole that allowed anyone to get a valid accessToken by calling the Cloud Function.